// apps/ripscloud-web/src/api/api.ts
export interface RipsAdminApplicationDTOsTenantDto {
    'id'?: string;
    'nit'?: string;
    'verificationDigit'?: string;
    'companyName'?: string;
    'commercialName'?: string;
    'taxRegime'?: string;
    'economicActivityCode'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'serviceCode'?: string;
    'logoUrl'?: string | null;
    'environment'?: number;
    'isActive'?: boolean;
    'createdAt'?: string;
    'updatedAt'?: string | null;
    'locations'?: Array<RipsAdminApplicationDTOsTenantLocationDto>;
}
export interface RipsAdminApplicationDTOsTenantLocationDto {
    'id'?: string;
    'name'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'habilitationCode'?: string;
    'isActive'?: boolean;
    'createdAt'?: string;
    'updatedAt'?: string | null;
}
export interface RipsAdminApplicationDTOsTenantLocationRequestDto {
    'name'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'habilitationCode'?: string;
    'isActive'?: boolean;
}
export interface RipsAdminApplicationDTOsTenantLocationUpdateRequestDto {
    'id'?: string | null;
    'name'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'habilitationCode'?: string;
    'isActive'?: boolean;
}
// ...
export interface RipsAdminApplicationDTOsUpdateTenantRequestDto {
    'tenantId'?: string;
    'nit'?: string;
    'verificationDigit'?: string;
    'companyName'?: string;
    'commercialName'?: string;
    'taxRegime'?: string;
    'economicActivityCode'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'serviceCode'?: string;
    'locations'?: Array<RipsAdminApplicationDTOsTenantLocationUpdateRequestDto>;
}